requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1